API service and map words

+++ app/src/main/java/com/islam/pureApp/domain/entites/Word.kt
@@ -1,3 +1,3 @@
 package com.islam.pureApp.domain.entites
 
-data class Word(val id: Long? = 0, val text: String, val count: Int)
+data class Word(val id: Int? = 0, val text: String, val count: Int)
